Modifications to parameter, telescope, and init files to add LDT/DeVeny
Following the example of the Bok/B&C Spectrograph addition, add parameters
for the Lowell Discovery Telescope / DeVeny Optical Spectrograph.  This is a
long-slit spectrograph, formerly known as the KPNO White Spectrograph (used on
the #1 36-inch Telescope), on long-term loan to Lowell Observatory.  It is
currently mounted on one of the cassegrain ports of the 4.3-m LDT near
Flagstaff, AZ.

	modified:   par/pypeitpar.py
	modified:   spectrographs/__init__.py
	modified:   telescopes.py