Clarify logic related to echelle & `ech_fix_format` in autoid.py
In the `ArchiveReid` class, there is logic related to whether an echelle
spectrograph is of fixed format or not.  Currently, this logic causes issues
with non-echelle spectrographs using the reidentify wavelength calibration
method because of how the default WavelengthSolutionPar keywords are defined.

Within pypeit.par.pypeitpar.WavelengthSolutionPar, the parameter `echelle`
defaults to `False`, while the parameter `ech_fix_format` defaults to `True`.
The idea is that when a spectrograph class specifies `echelle = True`, the
default behavior is for a fixed-format instrument.

The reidentify code in `ArchiveReid`, however, does not presently check whether
a spectrograph is indeed an echelle before executing code predicated upon
a fixed-format echelle spectrograph.

This commit offers one possible means for dealing with this logic.  It simply
adds a logical `AND`, checking whether the spectrograph is an echelle alongside
the check for a fixed format.

	modified:   pypeit/core/wavecal/autoid.py