Add LDT / DeVeny Spectrograph class
Following the example of the Bok/B&C Spectrograph addition, add parameters
for the Lowell Discovery Telescope / DeVeny Optical Spectrograph.  This is a
long-slit spectrograph, formerly known as the KPNO White Spectrograph (used on
the No.1 36-inch Telescope), on long-term loan to Lowell Observatory.  It is
currently mounted on one of the cassegrain ports of the 4.3-m LDT near
Flagstaff, AZ.

First pass at the class LDTDeVenySpectrograph() for addition to the PypeIt
repository.  Add the LDT/DeVeny to parameter, telescope, and init files, as
well as front-facing spectrograph lists.

This commit represents a thorough review of all PypeIt parameters, and making
adjustments based on the specific needs of the LDT/DeVeny spectrograph.

Overview:
For all calibration frames, use median combine rather than weighted mean
Specify the arc lamps available on DeVeny
Tweaks to some wavelength calibration values
Add fix for PypeIt issue #1155 (ech_fix_format = False)
Slit-edge setting for long-slit spectrographs
A couple default exposure times (bias/arc), otherwise there is no 'normal' range
A couple extraction and flux calibration adjustments

The parameters are now in the same order as shown in the online PypeIt
documentation for ease of reference.

	modified:   README.md
	modified:   doc/index.rst
	modified:   pypeit/par/pypeitpar.py
	modified:   pypeit/spectrographs/__init__.py
	new file:   pypeit/spectrographs/ldt_deveny.py
	modified:   pypeit/telescopes.py