Band-Aid to allow `reidentify` wavecal for longslit LDT/DeVeny
Unless / until PypeIt Issue #1155 is resolved, this change allows LDT/DeVeny
to use the `reidentify` wavecal method.  As discussed in the issue report,
in `pypeit/par/pyepitpar.py` two default wavecal parameters are set:
`echelle = False`, and `ech_fix_format = True`.  This means that if a
spectrograph class sets `echelle = True`, the fixed format parameter is, by
default, True.  The issue is that in pypeit/core/wavecal/autoid.py, there is
a check for echelle orders being specified if `ech_fix_format == True` (which
they are not for longslit spectrographs), but there is no parallel check to see
if the spectrograph is indeed an echelle.

This commit sets `ech_fix_format = False` for LDT/DeVeny so that the wavecal
method `reidentify` may be used with encountering this error.

If/when Issue #1155 is resolved to check for `echelle` along with
`ech_fix_format`, this parameter line may be removed.

	modified:   pypeit/spectrographs/ldt_deveny.py