Add shell for .show() method as part of pypeit_chk_tilts script
At present, the only means for checking the Tilts master is via the QA plots.
Other "pypeit_chk" scripts have a .show() method for plotting the Master in
ginga.  This commit adds a shell of such a method for the pypeit_chk_tilts
script.  The actual content of the method would still need to be written.

	modified:   pypeit/wavetilts.py