Add `wvrange` to WavelengthSolutionPar() to constrain `holy-grail`
While investigating the `holy-grail` method's failure to correctly identify
13 bright lines (with lamps `CdI`, `HgI`, `ArI`) in a long-slit spectrum
covering approximately 4500Å - 7500Å, I experimented with adding an additional
wavelength calibration parameter to ``WavelengthSolutionPar()`` to limit the
range of spectral lines from the ion linelists considered in the matching
algorithm.

This parameter is only used in the `holy-grail` method.  After the selected ion
linelists are read in and sorted, if this parameter is set, the collected
linelist is cropped to include only the wavelength range indicated.  This
cropped linelist is then fed into the pattern-matching algorithms as usual.

In testing, I had mixed results using this new parameter.  If I only specified
the `CdI` and `HgI` lamps, the wavelength calibration was done correctly -- but
only for the blue portion of this range since these metallic lamps do not have
lines redward of ~6000Å.  Including the `ArI` linelist in the matching caused
a failure of `holy-grail` to provide a wavelength solution.

I submit this commit if anyone wishes to use it as a starting point for
continued improvement of the `holy-grail` method.

	modified:   pypeit/core/wavecal/autoid.py
	modified:   pypeit/par/pypeitpar.py