Changes to pypeit_identify to use wvcalib.fits as MasterWaveCalib
For some LDT/DeVeny gratings and observation modes, `holy-grail` fails to
correctly identify the prominent lines in the spectrum.  This is partially due
do large slit widths (~3") used for some LDT observing programs.

It is possible to produce a workable wavelength solution manually with
`pypeit_identify`, but the resulting `wvcalib.fits` differed from the automated
`MasterWaveCalib_*.fits` that it could not be used to reduce Science frames.

This commit contains 2 changes:
  * Fixes 2 instances where 'slit' was used where 'spat_id' was needed.  For
    instance, in a DeVeny longslit observation, `slit=0` and `spat_id=247`.
    The `pypeit_identify` code was using `slit` where the main pipeline code
    uses `spat_id` in a couple of different places.
  * Compensates for the one place where the `strpar` JSON structure from
    `MasterWaveCalib_*.fits` is required to check whether this spectrograph is
    an echelle.  Since the use of manual wavelength identifications is likely
    restricted to spectrographs with continuously variable settings like DeVeny.

Changes to documentation for how to use a manual identification for pipeline
reduction forthcoming.

	modified:   pypeit/core/gui/identify.py
	modified:   pypeit/scripts/identify.py
	modified:   pypeit/wavecalib.py