Fixed Cen/Sky issue
This stems from an error in the NIHTS observation script for Cen/Sky nod
pattern that writes "Slit: None, Position:Cen" instead of "Slit: {slit},
Position:Sky" for the first COMMENT line.

In the compound_meta() method, the Sky frame is now listed as "Sky" in
the 'dithpos' metadata column, and the new "validate_fitstbl()" method
places the slit width from the associated "Cen" frame into the "Sky"
frame.

	modified:   pypeit/spectrographs/ldt_deveny.py
	modified:   pypeit/spectrographs/ldt_nihts.py