Add `set_calib_groups()` method
Add a new method `set_calib_groups()` to the Spectrograph class for the
purpose of automagically setting calibration groups *if defined by the
implementing spectrograph*.

The method in the base class raises a NotImplementedError rather than
simply returning the input metadata table because the calling routine
in `metadata.py` otherwise does some magic to set calibration groups
based on instrument setup.  This method is designed to be used for
specific spectrographs (*e.g.*, LDT/NIHTS) where calibration groups are
known *a priori* based on observing scripts.

	modified:   pypeit/metadata.py
	modified:   pypeit/spectrographs/spectrograph.py