Add edge tracing mods for stairstep slit mask
For stairstep slitmasks (e.g., LDT/NIHTS), the normal method of
defining slits by rise/fall (in the sobel mask) of intensity
doesn't work.  This commit adds an optional function in
`pypeit.core.trace` that adds new left edges to the right of
right edges and new right edges to the left of left edges.
In addition to the trace function, this adds a parameter to
the `EdgeTraceSet` that selects whether to call this function.

	modified:   pypeit/calibrations.py
	modified:   pypeit/core/trace.py
	modified:   pypeit/edgetrace.py
	modified:   pypeit/par/pypeitpar.py