Working but ugly hack to show multiple Spec1d
The current ``display.show_1dspec()`` function uses a static channel
for displaying spectra, causing repeated calls to ``pypeit_show_1dspec``
to overwrite the existing spectrum.

This commit reads in the list of current channels and creates a new
channel name if the default is already taken.

This should be finessed to allow a command-line argument to
``pypeit_show_1dspec`` for the user to specify whether or not to
create a new channel.  Will get there...

	modified:   pypeit/display/display.py